I love this tool!  and here is gentle nudge
github makes it super easy to change from using the master nomenclature to 'main' instead.  In fact, it's been the default for 'main' to be the main branch for repos for a while now.  The instructions to migrate old repos are pretty pain free https://github.com/github/renaming .Even if you don't choose to, I think it does, or soon will be the case potential users would expect to see 'main' used in docs relevant to how to install in their own repos. 

Thank you--
jem